+++ TopGame/TopGame/entity.cpp
@@ -8,10 +8,10 @@ Entity::Entity(Image &image, String Name, float X, float Y, int W, int H)
 	life = true; onGround = false;
 	texture.loadFromImage(image);
 	sprite.setTexture(texture);
-	sprite.setOrigin(w / 2, h / 2);
+	sprite.setOrigin((float)w / 2.f, (float)h / 2.f);
 }
 
 FloatRect Entity::GetRect()
 {
-	return FloatRect(x, y, w, h);
+	return FloatRect((float)x, (float)y, (float)w, (float)h);
 }
